Organize file structure
“Organizing ahead of time makes the work more enjoyable” …and avoids
nightmares.
Create asset folder to contain graphical elements and application
folder for each game or interactive work of art.
Create folder for the lava detail and move vectors to it.
Move Lava.pde to the sketch folder that uses it; files and their dependancies must be in their respective sketch folder.
Rename TEST for significance.
Edit svg references in to accommodate new file structure.